Alias mejorado y separacion de composables
