+++ app/src/main/java/com/example/tictactoe/inicio/Inicio.kt
@@ -1,0 +1,51 @@
+package com.example.tictactoe.inicio
+
+import android.annotation.SuppressLint
+import android.app.Activity
+import android.content.Context
+import android.content.res.Configuration
+import android.widget.Toast
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import com.example.tictactoe.R
+
+@SuppressLint("ContextCastToActivity")
+@Composable
+fun Inicio(
+    navController: NavController,
+    primerJuego: Boolean
+){
+    val configuration = LocalConfiguration.current
+    when (configuration.orientation) {
+        Configuration.ORIENTATION_PORTRAIT -> {
+            InicioPortrait(navController, primerJuego)
+        }
+        Configuration.ORIENTATION_LANDSCAPE -> {
+            InicioLandscape(navController, primerJuego)
+        }
+        else -> {
+            InicioPortrait(navController, primerJuego)
+        }
+    }
+}
+
+
