+++ app/src/main/java/com/example/tictactoe/instrucciones/Instrucciones.kt
@@ -1,0 +1,36 @@
+package com.example.tictactoe.instrucciones
+
+import android.content.res.Configuration
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import com.example.tictactoe.R
+
+@Composable
+fun Instrucciones(navController: NavController) {
+    val configuration = LocalConfiguration.current
+    when (configuration.orientation) {
+        Configuration.ORIENTATION_PORTRAIT -> {
+            InstruccionesPortrait(navController)
+        }
+        Configuration.ORIENTATION_LANDSCAPE -> {
+            InstruccionesLandscape(navController)
+        }
+        else -> {
+            InstruccionesPortrait(navController)
+        }
+    }
+}
