+++ app/src/main/java/com/example/tictactoe/instrucciones/InstruccionesLandscape.kt
@@ -1,51 +1,19 @@
-package com.example.tictactoe.screens
+package com.example.tictactoe.instrucciones
 
-import android.content.res.Configuration
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.example.tictactoe.R
-
-@Composable
-fun Instrucciones(navController: NavController) {
-    val configuration = LocalConfiguration.current
-    when (configuration.orientation) {
-        Configuration.ORIENTATION_PORTRAIT -> {
-            InstruccionesPortrait(navController)
-        }
-        Configuration.ORIENTATION_LANDSCAPE -> {
-            InstruccionesLandscape(navController)
-        }
-        else -> {
-            InstruccionesPortrait(navController)
-        }
-    }
-}
-
-@Composable
-fun InstruccionesPortrait(navController: NavController) {
-    val scrollState = rememberScrollState()
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .verticalScroll(scrollState)
-    ) {
-        InstruccionesTexts()
-    }
-}
 
 @Composable
 fun InstruccionesLandscape(navController: NavController) {
@@ -89,38 +57,6 @@ fun InstruccionesLandscape(navController: NavController) {
                     style = TextStyle(textAlign = TextAlign.Justify)
                 )
             }
-        }
-    }
-}
-
-@Composable
-fun InstruccionesTexts() {
-    Column(modifier = Modifier.padding(16.dp)) {
-        Row(modifier = Modifier.padding(bottom = 16.dp)) {
-            Text(
-                text = stringResource(R.string.instrucciones),
-                style = TextStyle(textAlign = TextAlign.Justify)
-            )
-        }
-
-        Row(modifier = Modifier.padding(bottom = 16.dp)) {
-            Text(
-                text = stringResource(R.string.instrucciones_perfil),
-                style = TextStyle(textAlign = TextAlign.Justify)
-            )
-        }
-        Row(modifier = Modifier.padding(bottom = 16.dp)) {
-            Text(
-                text = stringResource(R.string.instrucciones_dificultad),
-                style = TextStyle(textAlign = TextAlign.Justify)
-            )
-        }
-
-        Row {
-            Text(
-                text = stringResource(R.string.instrucciones_navegacion),
-                style = TextStyle(textAlign = TextAlign.Justify)
-            )
         }
     }
 }
