+++ app/src/main/java/com/example/tictactoe/instrucciones/InstruccionesPortrait.kt
@@ -1,0 +1,21 @@
+package com.example.tictactoe.instrucciones
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.navigation.NavController
+
+@Composable
+fun InstruccionesPortrait(navController: NavController) {
+    val scrollState = rememberScrollState()
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .verticalScroll(scrollState)
+    ) {
+        InstruccionesTexts()
+    }
+}
