+++ app/src/main/res/values/strings.xml
@@ -50,5 +50,6 @@
     <string name="tiempo_restante_r">Tiempo Restante: %1$02d:%2$02d</string>
     <string name="toast_tiempo_cero">El tiempo máximo no puede ser 00:00.</string>
     <string name="resultados">Resultados de la partida:</string>
+    <string name="toast_alias_vacio">El alías no puede estar vacío.</string>
 
 </resources>
